+++ benchmarks/aeson-benchmarks.cabal
@@ -15,28 +15,30 @@ library
 
   exposed-modules:
     Data.Aeson
+    Data.Aeson.Encode
     Data.Aeson.Encoding
     Data.Aeson.Encoding.Internal
-    Data.Aeson.Encode
-    Data.Aeson.Encode.Builder
-    Data.Aeson.Encode.Functions
-    Data.Aeson.Functions
     Data.Aeson.Internal
     Data.Aeson.Internal.Time
     Data.Aeson.Parser
+    Data.Aeson.Types
+    Data.Aeson.TH
+
+    Data.Aeson.Encode.Builder
+    Data.Aeson.Functions
     Data.Aeson.Parser.Internal
     Data.Aeson.Parser.Time
-    Data.Aeson.TH
-    Data.Aeson.Types
-    Data.Aeson.Types.Class
+    Data.Aeson.Types.FromJSON
     Data.Aeson.Types.Generic
-    Data.Aeson.Types.Instances
+    Data.Aeson.Types.ToJSON
+    Data.Aeson.Types.Class
     Data.Aeson.Types.Internal
-    Data.Aeson.Types.Instances.Tuple
 
   build-depends:
     attoparsec >= 0.13.0.1,
     base == 4.*,
+    base-compat >= 0.9.1 && <0.10,
+    time-locale-compat >=0.1.1 && <0.2,
     bytestring >= 0.10.4.0,
     containers,
     deepseq,
